requirements.txt
beautifulsoup4
requests
neologdn
pymongo
mysql-connector-python
